modified quickflat.make() to properly display pixels outside of cortex as transparent

+++ cortex/quickflat.py
@@ -425,9 +425,9 @@ def make(braindata, height=1024, recache=False, **kwargs):
         return img.transpose(1,0,2)[::-1], extents
     else:
         badmask = np.array(pixmap.sum(1) > 0).ravel()
-        img = (np.nan*np.ones(mask.shape)).astype(braindata.data.dtype)
-        mimg = (np.nan*np.ones(badmask.shape)).astype(braindata.data.dtype)
-        mimg[badmask] = (pixmap*data.ravel())[badmask].astype(mimg.dtype)
+        img = (np.nan*np.ones(mask.shape))
+        mimg = (np.nan*np.ones(badmask.shape))
+        mimg[badmask] = (pixmap*data.ravel())[badmask]
         img[mask] = mimg
 
         return img.T[::-1], extents
